+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Update horizontal padding of overdue section header
 - Link overdue list adapter to the `RecyclerView` in `OverdueScreen`
 - Implement `OverdueAppointmentListItemNew` adapter
+- Rename `SeeAllPendingAppointmentList` to `PendingListFooter`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 - [In Progress: 03 Jun 2022] Implement `OverdueAppointmentListItemNew` adapter